fix: reducce useConfig singleton usage
