+++ packages/server-core/src/adapters/http/controllers/entities/identity-provider/index.ts
@@ -5,6 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './factory';
 export * from './handlers';
 export * from './module';
