+++ packages/server-core/src/adapters/http/controllers/index.ts
@@ -7,4 +7,3 @@
 
 export * from './entities';
 export * from './workflows';
-export * from './register';
