+++ packages/server-core/src/adapters/http/controllers/workflows/authorize/index.ts
@@ -5,6 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './factory';
 export * from './module';
 export * from './types';
