+++ packages/server-core/src/adapters/http/controllers/workflows/token/index.ts
@@ -6,4 +6,3 @@
  */
 
 export * from './module';
-export * from './factory';
