+++ packages/server-core/src/adapters/http/middleware/built-in/rate-limit.ts
@@ -5,40 +5,30 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { REALM_MASTER_NAME } from '@authup/core-kit';
 import type { OptionsInput } from '@routup/rate-limit';
 import { rateLimit } from '@routup/rate-limit';
 import type { Request, Router } from 'routup';
 import { merge } from 'smob';
-import { useConfig } from '../../../../config';
 import { useRequestIdentity } from '../../request';
 
 export function registerRateLimitMiddleware(router: Router, input?: OptionsInput) {
-    const config = useConfig();
     let options : OptionsInput = {
-        skip(req: Request) {
-            const identity = useRequestIdentity(req);
-
-            return !!identity &&
-                identity.type === 'robot' &&
-                identity.realmName === REALM_MASTER_NAME &&
-                !!identity.data &&
-                identity.data.name === config.robotAdminName;
-        },
         max(req: Request) {
             const identity = useRequestIdentity(req);
-            if (identity && identity.type === 'user') {
-                return 60 * 100; // 100 req p. sec
+            if (!identity) {
+                return 60 * 20; // 20 req. p. sec
             }
 
-            if (
-                identity &&
-                (identity.type === 'robot' || identity.type === 'client')
-            ) {
-                return 60 * 1000; // 1000 req p. sec
+            switch (identity.type) {
+                case 'robot':
+                case 'client': {
+                    return 60 * 5_000; // 1000 req p. sec
+                }
+                case 'user':
+                default: {
+                    return 60 * 100; // 100 req p. sec
+                }
             }
-
-            return 60 * 20; // 20 req p. sec
         },
         windowMs: 60 * 1000, // 60 sec
     };
