+++ packages/server-core/src/app/modules/config/module.ts
@@ -8,8 +8,8 @@
 import type { Module } from '../types';
 import { ConfigInjectionKey } from './constants';
 import type { Config, ConfigRawReadOptions } from '../../../config';
-import { normalizeConfig, readConfigRaw } from '../../../config';
-import type { IDIContainer } from '../../../core/di/types';
+import { normalizeConfig, readConfigRaw, setConfig } from '../../../config';
+import type { IDIContainer } from '../../../core';
 
 export class ConfigModule implements Module {
     protected instance : Config | undefined;
@@ -29,6 +29,8 @@ export class ConfigModule implements Module {
         } else {
             instance = await this.read();
         }
+
+        setConfig(instance);
 
         container.register(ConfigInjectionKey, {
             useValue: instance,
