--- packages/server-core/src/core/identity/constants.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2025.
- *  Author Peter Placzek (tada5hi)
- *  For the full copyright and license information,
- *  view the LICENSE file that was distributed with this source code.
- */
-
-export const IDENTITY_PROVIDER_ACCOUNT_MANAGER_TOKEN = Symbol('IdentityProviderAccountManager');
-
-export const IDENTITY_RESOLVER_TOKEN = Symbol('IdentityResolver');
