+++ packages/server-core/src/core/identity/index.ts
@@ -8,5 +8,3 @@
 export * from './entities';
 export * from './provider';
 export * from './resolver';
-
-export * from './constants';
