--- packages/server-core/src/core/ldap/constants.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2025.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-export const LDAP_CLIENT_FACTORY_TOKEN = Symbol('LdapClientFactory');
