+++ packages/server-core/src/core/ldap/index.ts
@@ -5,5 +5,4 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './constants';
 export * from './types';
