+++ packages/server-core/src/index.ts
@@ -7,11 +7,11 @@
 
 import 'reflect-metadata';
 
+export * from './app';
 export * from './cli/commands';
 export * from './config';
 export * from './adapters/database/domains';
 export * from './adapters/database';
 export * from './adapters/http';
-export * from './app/module';
 export * from './security';
 export * from './services';
